teledocker/requirements.txt
python-telegram-bot>=20.0
docker>=6.0.0
pymongo>=4.3.3
redis>=4.5.4
python-dotenv>=1.0.0